# docs/swagger.yaml
definitions:
  models.Bucket:
    properties:
      access_key:
        type: string
      created_at:
        type: string
      custom_config:
        type: string
      endpoint:
        type: string
      id:
        type: string
      name:
        type: string
      region:
        type: string
      s3_provider:
        type: string
      secret_key:
        type: string
      storage_type:
        $ref: '#/definitions/models.StorageType'
      updated_at:
        type: string
      use_ssl:
        type: boolean
    required:
    - access_key
    - endpoint
    - name
    - region
    - secret_key
    type: object
...
  models.StorageType:
    enum:
    - 0
    - 1
    format: int32
    type: integer
    x-enum-varnames:
    - HotStorage
    - ColdStorage
...
  router.BucketInsertDTO:
    properties:
      access_key:
        type: string
      custom_config:
        type: string
      endpoint:
        type: string
      name:
        type: string
      region:
        type: string
      s3_provider:
        type: string
      secret_key:
        type: string
      storage_type:
        $ref: '#/definitions/models.StorageType'
      use_ssl:
        type: boolean
    required:
    - access_key
    - endpoint
    - name
    - region
    - secret_key
    type: object
...
  router.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: error message
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/v1/bucket/:
    get:
      description: List all S3 buckets
      operationId: ListBuckets
      parameters:
      - description: API Token
        in: header
        name: x-api-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Bucket'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: List buckets
      tags:
      - buckets
    post:
      consumes:
      - application/json
      description: Create a new S3 bucket. Only admin users can create buckets.
      operationId: CreateBucket
      parameters:
      - description: API Token
        in: header
        name: x-api-token
        required: true
        type: string
      - description: Bucket
        in: body
        name: bucket
        required: true
        schema:
          $ref: '#/definitions/router.BucketInsertDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Bucket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Create bucket
      tags:
      - buckets
  /api/v1/bucket/{id}:
    delete:
      description: Delete a bucket by ID. Only admin users can delete buckets.
      operationId: DeleteBucket
      parameters:
      - description: API Token
        in: header
        name: x-api-token
        required: true
        type: string
      - description: Bucket ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Delete bucket
      tags:
      - buckets
    get:
      description: Get a bucket by ID
      operationId: GetBucket
      parameters:
      - description: API Token
        in: header
        name: x-api-token
        required: true
        type: string
      - description: Bucket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Bucket'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Get bucket
      tags:
      - buckets
    patch:
      consumes:
      - application/json
      description: Update a bucket by ID. Only admin users can update buckets.
      operationId: UpdateBucket
      parameters:
      - description: API Token
        in: header
        name: x-api-token
        required: true
        type: string
      - description: Bucket ID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket
        in: body
        name: bucket
        required: true
        schema:
          $ref: '#/definitions/router.BucketInsertDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Bucket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Update bucket
      tags:
      - buckets
...
swagger: "2.0"